+++ lib/matching/src/AccessFunction.cpp
@@ -70,7 +70,7 @@ llvm::ArrayRef<SingleDimensionAccess> AccessFunction::getDimensionAccesses() con
 	return functions;
 }
 
-void AccessFunction::map(llvm::SmallVectorImpl<size_t>& results, llvm::ArrayRef<long> equationIndexes) const
+void AccessFunction::map(llvm::SmallVectorImpl<long>& results, llvm::ArrayRef<long> equationIndexes) const
 {
 	for (const auto& function : functions)
 		results.push_back(function(equationIndexes));
